# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a DEIP server API.",
    "version": "1.0.0",
    "title": "DEIP",
    "contact": {
      "email": "yahor.tsaryk@deip.world"
    }
  },
...
  "paths": {
...
    "/api/v2/proposals/{proposalId}": {
      "get": {
        "tags": [
          "proposal"
        ],
        "summary": "Find proposal",
        "description": "Return a proposal",
        "operationId": "getProposalById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "description": "ID of project to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Proposal"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Proposal not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v2/attributes": {
      "get": {
        "tags": [
          "attribute"
        ],
        "summary": "Find all attributes",
        "description": "Return an array of attributes",
        "operationId": "getAttributes",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Attribute"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Attributes not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v2/attributes/scope/{scope}": {
      "get": {
        "tags": [
          "attribute"
        ],
        "summary": "Find attributes",
        "description": "Return an array of attributes",
        "operationId": "getAttributesByScope",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "description": "scope of attributes",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Attribute"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Attributes not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v2/attribute/{id}": {
      "get": {
        "tags": [
          "attribute"
        ],
        "summary": "Find attribute",
        "description": "Return an attribute",
        "operationId": "getAttributeById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of attribute",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Attribute"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Attribute not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v2/attribute": {
      "post": {
        "tags": [
          "attribute"
        ],
        "summary": "Create attribute",
        "description": "",
        "operationId": "attributesService.create",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "isPublished": {
                  "type": "boolean",
                  "default": false
                },
                "isFilterable": {
                  "type": "boolean",
                  "default": false
                },
                "isHidden": {
                  "type": "boolean",
                  "default": false
                },
                "isMultiple": {
                  "type": "boolean",
                  "default": false
                },
                "title": {
                  "type": "string"
                },
                "shortTitle": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "valueOptions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "shortTitle": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "value": {
                        "type": "string"
                      }
                    }
                  }
                },
                "defaultValue": {},
                "schemas": {
                  "type": "object"
                },
                "isEditable": {
                  "type": "boolean"
                },
                "isRequired": {
                  "type": "boolean",
                  "default": false
                },
                "scope": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "attribute"
        ],
        "summary": "Update attribute",
        "description": "",
        "operationId": "attributesService.update",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "isPublished": {
                  "type": "boolean",
                  "default": false
                },
                "isFilterable": {
                  "type": "boolean",
                  "default": false
                },
                "isHidden": {
                  "type": "boolean",
                  "default": false
                },
                "isMultiple": {
                  "type": "boolean",
                  "default": false
                },
                "title": {
                  "type": "string"
                },
                "shortTitle": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "valueOptions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "shortTitle": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "value": {
                        "type": "string"
                      }
                    }
                  }
                },
                "defaultValue": {},
                "schemas": {
                  "type": "object"
                },
                "isEditable": {
                  "type": "boolean"
                },
                "isRequired": {
                  "type": "boolean",
                  "default": false
                },
                "isSystem": {
                  "type": "boolean",
                  "default": false
                },
                "scope": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/attribute/file/{scope}/{_id}/{attributeId}/{filename}": {
      "get": {
        "tags": [
          "attribute"
        ],
        "summary": "Find attribute",
        "description": "Return an attribute",
        "operationId": "getAttributeFile",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "description": "scope of attribute",
            "required": true,
            "type": "string"
          },
          {
            "name": "_id",
            "in": "path",
            "description": "id of entity",
            "required": true,
            "type": "string"
          },
          {
            "name": "attributeId",
            "in": "path",
            "description": "id of attribute",
            "required": true,
            "type": "string"
          },
          {
            "name": "filename",
            "in": "path",
            "description": "name of file",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Attribute"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Attribute not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Proposal": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "portalId": {
          "type": "string"
        },
        "cmd": {
          "type": "object",
          "properties": {
            "CMD_NUM": {
              "type": "number"
            },
            "CMD_PAYLOAD": {
              "type": "string"
            }
          }
        },
        "proposer": {
          "type": "string"
        },
        "parties": {
          "type": "object",
          "properties": {
            "isProposer": {
              "type": "boolean"
            },
            "status": {
              "type": "number"
            },
            "account": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                }
              }
            },
            "isHidden": {
              "type": "boolean"
            },
            "signers": {
              "type": "object",
              "properties": {
                "signer": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    }
                  }
                },
                "txInfo": {}
              }
            }
          }
        },
        "type": {
          "type": "number"
        },
        "status": {
          "type": "number"
        },
        "decisionMakers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "approvers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "rejectors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "Attribute": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "portalId": {
          "type": "string"
        },
        "isSystem": {
          "type": "boolean"
        },
        "type": {
          "type": "string"
        },
        "isFilterable": {
          "type": "boolean"
        },
        "isEditable": {
          "type": "boolean"
        },
        "isRequired": {
          "type": "boolean"
        },
        "isHidden": {
          "type": "boolean"
        },
        "isMultiple": {
          "type": "boolean"
        },
        "title": {
          "type": "string"
        },
        "shortTitle": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "schemas": {
          "type": "object"
        },
        "valueOptions": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "shortTitle": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "value": {
                "type": "string"
              }
            }
          }
        },
        "defaultValue": {},
        "blockchainFieldMeta": {
          "type": "object",
          "properties": {
            "field": {
              "type": "string"
            },
            "isPartial": {
              "type": "boolean"
            }
          }
        },
        "scope": {
          "type": "string"
        },
        "isGlobalScope": {
          "type": "boolean"
        }
      }
    },
...
  }
}